+++ packages/liveblocks-react-ui/src/components/Composer.tsx
@@ -61,12 +61,12 @@ import type {
   ComposerEditorMentionSuggestionsProps,
   ComposerEditorProps,
   ComposerFormProps,
+  ComposerMarkToggleProps,
   ComposerSubmitComment,
-  ComposerTextFormatToggleProps,
 } from "../primitives/Composer/types";
 import { useComposerAttachmentsDropArea } from "../primitives/Composer/utils";
 import { MENTION_CHARACTER } from "../slate/plugins/mentions";
-import type { ComposerBodyTextFormat } from "../types";
+import type { ComposerBodyMark } from "../types";
 import { classNames } from "../utils/class-names";
 import { useControllableState } from "../utils/use-controllable-state";
 import { useLayoutEffect } from "../utils/use-layout-effect";
@@ -93,7 +93,7 @@ interface EmojiEditorActionProps extends EditorActionProps {
   onPickerOpenChange?: EmojiPickerProps["onOpenChange"];
 }
 
-interface TextFormatToggleProps extends ComposerTextFormatToggleProps {
+interface MarkToggleProps extends ComposerMarkToggleProps {
   shortcut?: ReactNode;
 }
 
@@ -177,6 +177,11 @@ export type ComposerProps<M extends BaseMetadata = DM> = Omit<
     showAttachments?: boolean;
 
     /**
+     * Whether to show formatting controls (e.g. a floating toolbar with formatting toggles when selecting text)
+     */
+    showFormattingControls?: boolean;
+
+    /**
      * Whether the composer is disabled.
      */
     disabled?: ComposerFormProps["disabled"];
@@ -212,6 +217,7 @@ interface ComposerEditorContainerProps
     ComposerProps,
     | "defaultValue"
     | "showAttachments"
+    | "showFormattingControls"
     | "showAttribution"
     | "overrides"
     | "actions"
@@ -368,16 +374,11 @@ function ComposerMentionSuggestions({
   ) : null;
 }
 
-function TextFormatToggle({
-  format,
-  shortcut,
-  children,
-  ...props
-}: TextFormatToggleProps) {
+function MarkToggle({ mark, shortcut, children, ...props }: MarkToggleProps) {
   const $ = useOverrides();
   const label = useMemo(() => {
-    return $.COMPOSER_TOGGLE_TEXT_FORMAT(format);
-  }, [$, format]);
+    return $.COMPOSER_TOGGLE_MARK(mark);
+  }, [$, mark]);
 
   return (
     <ShortcutTooltip
@@ -385,21 +386,23 @@ function TextFormatToggle({
       shortcut={shortcut}
       sideOffset={FLOATING_ELEMENT_SIDE_OFFSET + 2}
     >
-      <ComposerPrimitive.TextFormatToggle format={format} asChild {...props}>
-        <Button aria-label={label}>{children}</Button>
-      </ComposerPrimitive.TextFormatToggle>
+      <ComposerPrimitive.MarkToggle mark={mark} asChild {...props}>
+        <Button aria-label={label} variant="toggle">
+          {children}
+        </Button>
+      </ComposerPrimitive.MarkToggle>
     </ShortcutTooltip>
   );
 }
 
-type TextFormatToggles = {
-  [K in ComposerBodyTextFormat]: ComponentType<PropsWithChildren>;
+type MarkToggles = {
+  [K in ComposerBodyMark]: ComponentType<PropsWithChildren>;
 };
 
-const textFormatToggles: TextFormatToggles = {
+const markToggles: MarkToggles = {
   bold: () => (
-    <TextFormatToggle
-      format="bold"
+    <MarkToggle
+      mark="bold"
       shortcut={
         <>
           <ShortcutTooltipKey name="mod" />
@@ -408,11 +411,11 @@ const textFormatToggles: TextFormatToggles = {
       }
     >
       <BoldIcon />
-    </TextFormatToggle>
+    </MarkToggle>
   ),
   italic: () => (
-    <TextFormatToggle
-      format="italic"
+    <MarkToggle
+      mark="italic"
       shortcut={
         <>
           <ShortcutTooltipKey name="mod" />
@@ -421,11 +424,11 @@ const textFormatToggles: TextFormatToggles = {
       }
     >
       <ItalicIcon />
-    </TextFormatToggle>
+    </MarkToggle>
   ),
   strikethrough: () => (
-    <TextFormatToggle
-      format="strikethrough"
+    <MarkToggle
+      mark="strikethrough"
       shortcut={
         <>
           <ShortcutTooltipKey name="mod" />
@@ -435,11 +438,11 @@ const textFormatToggles: TextFormatToggles = {
       }
     >
       <StrikethroughIcon />
-    </TextFormatToggle>
+    </MarkToggle>
   ),
   code: () => (
-    <TextFormatToggle
-      format="code"
+    <MarkToggle
+      mark="code"
       shortcut={
         <>
           <ShortcutTooltipKey name="mod" />
@@ -448,18 +451,18 @@ const textFormatToggles: TextFormatToggles = {
       }
     >
       <CodeIcon />
-    </TextFormatToggle>
+    </MarkToggle>
   ),
 };
 
-const textFormatTogglesList = Object.entries(textFormatToggles).map(
-  ([format, Toggle]) => <Toggle key={format} />
-);
+const markTogglesList = Object.entries(markToggles).map(([mark, Toggle]) => (
+  <Toggle key={mark} />
+));
 
 function ComposerFloatingToolbar() {
   return (
     <ComposerPrimitive.FloatingToolbar className="lb-root lb-portal lb-elevation lb-composer-floating-toolbar">
-      {textFormatTogglesList}
+      {markTogglesList}
     </ComposerPrimitive.FloatingToolbar>
   );
 }
@@ -538,7 +541,7 @@ function ComposerAttachments({
   );
 }
 
-const editorComponents: ComposerEditorComponents = {
+const editorRequiredComponents: ComposerEditorComponents = {
   Mention: ComposerMention,
   MentionSuggestions: ComposerMentionSuggestions,
   FloatingToolbar: ComposerFloatingToolbar,
@@ -547,6 +550,7 @@ const editorComponents: ComposerEditorComponents = {
 
 function ComposerEditorContainer({
   showAttachments = true,
+  showFormattingControls = true,
   showAttribution,
   defaultValue,
   isCollapsed,
@@ -562,6 +566,14 @@ function ComposerEditorContainer({
   const { isEmpty } = useComposer();
   const { hasMaxAttachments } = useComposerAttachmentsContext();
   const $ = useOverrides(overrides);
+  const components = useMemo(() => {
+    return {
+      ...editorRequiredComponents,
+      FloatingToolbar: showFormattingControls
+        ? ComposerFloatingToolbar
+        : undefined,
+    };
+  }, [showFormattingControls]);
 
   const [isDraggingOver, dropAreaProps] = useComposerAttachmentsDropArea({
     disabled: disabled || hasMaxAttachments,
@@ -587,7 +599,7 @@ function ComposerEditorContainer({
         placeholder={$.COMPOSER_PLACEHOLDER}
         defaultValue={defaultValue}
         autoFocus={autoFocus}
-        components={editorComponents}
+        components={components}
         disabled={disabled}
         dir={$.dir}
       />
@@ -678,6 +690,7 @@ export const Composer = forwardRef(
       autoFocus,
       disabled,
       showAttachments = true,
+      showFormattingControls = true,
       showAttribution,
       roomId: _roomId,
       ...props
@@ -853,6 +866,7 @@ export const Composer = forwardRef(
             isCollapsed={isCollapsed}
             showAttachments={showAttachments}
             showAttribution={showAttribution}
+            showFormattingControls={showFormattingControls}
             hasResolveMentionSuggestions={hasResolveMentionSuggestions}
             onEmptyChange={setEmptyRef}
             onEmojiPickerOpenChange={setEmojiPickerOpenRef}