+++ packages/liveblocks-react-ui/src/components/internal/Button.tsx
@@ -6,7 +6,7 @@ import React, { forwardRef } from "react";
 import { classNames } from "../../utils/class-names";
 
 export interface ButtonProps extends ComponentProps<"button"> {
-  variant?: "default" | "outline" | "primary";
+  variant?: "default" | "outline" | "toggle" | "primary";
   size?: "default" | "large";
   disableable?: boolean;
 }